--- src/FluentImposter.DataStore.AwsDynamoDb/DataStoreExtension.cs
@@ -1,17 +1,0 @@
-﻿using Amazon.DynamoDBv2;
-
-using FluentImposter.Core;
-
-namespace FluentImposter.DataStore.AwsDynamoDb
-{
-    public static class DataStoreExtension
-    {
-        public static ImpostersAsMockConfiguration UseDynamoDb(this ImpostersAsMockConfiguration impostersConfiguration,
-                                                        IAmazonDynamoDB amazonDynamoDbClient)
-        {
-            impostersConfiguration.SetDataStore(new AwsDynamoDbDataStore(amazonDynamoDbClient));
-
-            return impostersConfiguration;
-        }
-    }
-}
