+++ tests/FluentImposter.AspnetCore.Tests.Integration/FluentImposter.AspnetCore.Tests.Integration.csproj
@@ -244,7 +244,6 @@
   </ItemGroup>
   <ItemGroup>
     <Compile Include="ApplicationBuilderExtensionTests.cs" />
-    <Compile Include="Fakes\FakeDataStoreExtension.cs" />
     <Compile Include="Fakes\FakeImposterStubbingAResourceWithAPatternInUri.cs" />
     <Compile Include="Fakes\FakeImposterWithMockedRequestContent.cs" />
     <Compile Include="Fakes\FakeImposterWithRequestContent.cs" />
